third_party/blink/web_tests/external/wpt/css/cssom-view/client-props-input-expected.txt

This is a testharness.js-based test.
[FAIL] client* on input / textarea
  assert_equals: INPUT : clientWidth should be the padding box for textarea, content box for input expected 300 but got 304
Harness: the test ran to completion.
